Add a benchmark and the codspeed integration
