+++ .github/workflows/ci.yml
@@ -40,6 +40,12 @@ jobs:
       - name: Run pytest
         uses: pavelzw/pytest-action@510c5e90c360a185039bea56ce8b3e7e51a16507 # v2.2.0
 
+      - name: Run benchmarks
+        uses: CodSpeedHQ/action@513a19673a831f139e8717bf45ead67e47f00044 # v3.2.0
+        with:
+          token: ${{ secrets.CODSPEED_TOKEN }}
+          run: pytest --codspeed
+
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
