+++ pyproject.toml
@@ -57,6 +57,7 @@ test = [
   "coverage[toml]>=7.6.3",
   "gitpython>=3.1.43",
   "pytest>=8.3.2",
+  "pytest-codspeed>=3.1.2",
   "pytest-cov>=5.0.0",
   "pytest-emoji>=0.2.0",
   "pytest-md>=0.2.0",
